Add files via upload
easypages.js is a node module for streamlining the creation of simple web applications.  It allows for each page to be created with a single line of code, along with a single line as a header and another as a footer.  It currently supports absolute URL's and ostensibly also supports RegExp URL's (though this has not yet been tested).  Future commits will refine this capability.  Supported file types include HTML, CSS, JavaScript, PNG, JPEG, GIF and BMP.
The compressed folder demo.zip showcases a sample implementation of the easypages.js module.  It is based on an assignment from CodingDojo Programming Bootcamp.  All images in the 'images' folder are the property of their respective owners.  Special thanks to Jack Elkan for assistance in the selection of car pictures.